pyproject.toml
[project]
name = "gemini-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-auth>=2.48.0",
    "google-auth-oauthlib>=1.2.1",
    "google-generativeai>=0.8.6",
]
